Export EIP-6963 Types
